Move Priority enum to _constants.py
Move the Priority enum class from _pswindows.py to _constants.py. Priority
enum is now defined once in _constants and imported by _pswindows, simplifying
the code organization.
